Extract waveform state interaction
